crud_done

+++ online_judge/frontend/src/assets/hs/problemset/problemset.jsx
@@ -4,15 +4,15 @@ import '../problemset/problemset.css'
 import {fetchDataFromDatabase} from './apilinkps.js'
 import { MdDelete,MdOutlineUpdate } from "react-icons/md";
 import { deleteDataFromDatabase } from './delprob_api.js';
-import { Link } from "react-router-dom";
+import { useNavigate,Link } from "react-router-dom";
 
 export const Problemset = ()=>{
     const curr_mode = localStorage.getItem('curr-theme');
     const [details,setDetails] = useState([]);
     const [state,setState] = useState(curr_mode?curr_mode:'lightmode');
     const [kolor,setKolor] = useState('');
-    const [delres,setDelres] = useState('');
     const [bc,setBc] = useState('');
+    const navigate = useNavigate();
 
     useEffect(()=>{
         localStorage.setItem('curr-theme',state);
@@ -27,9 +27,18 @@ export const Problemset = ()=>{
     },[])
     console.log('lol',details);
 
-    const eventDelete = async (name)=>{
-        const response = await deleteDataFromDatabase(name);
-        setDelres(response);
+    const handleDelete = async (name)=>{
+        if(confirm('Please confirm deletion'))
+        {
+            const response = await deleteDataFromDatabase(name);
+            //reload after deleting response
+            window.location.reload();
+        }
+    }
+
+    const sendName = (name)=>{
+        localStorage.setItem('sendUpdateName',name);
+        navigate('/updaterecord');
     }
 
     return(
@@ -38,7 +47,7 @@ export const Problemset = ()=>{
             <h1 className={`${kolor} heamder`}>Problemset</h1>
             <table width="100%" className={`tablecover ${bc}`}>
                 <thead>
-                    <tr className={`displaytable ${kolor}`} align="center">
+                    <tr className={`displaytable`} align="center">
                         <th width="5%"  height="40px" className={`${bc}`}>S.No.</th>
                         <th width="25%" height="40px" className={`${bc}`}>Problem Title</th>
                         <th width="10%" height="40px" className={`${bc}`}>Problem Status</th>
@@ -50,16 +59,16 @@ export const Problemset = ()=>{
                 </thead>
                 <tbody>
                     {
-                        details.map((entry,i)=>{
+                        details && details.map((entry,i)=>{
                             return(
                                 <tr key={i} align="center" className={`${kolor} editing`}>
                                     <td className={`${bc}`}>{i+1}</td>
-                                    <td className={`${bc}`}>{entry.problem_name}</td>
+                                    <td className={`${bc}`}><Link to={`/displayprob/${entry._id}`} className="nodecor">{entry.problem_name}</Link></td>
                                     <td className={`${bc}`}>{entry.problem_status}</td>
                                     <td className={`${bc}`}>{entry.difficulty}</td>
                                     <td className={`${bc}`}>{entry.tags.join(',')}</td>
-                                    <td className={`${bc}`}><Link to="/updaterecord"><MdOutlineUpdate className='botton'/></Link></td>
-                                    <td className={`${bc}`}><MdDelete className='botton'/>{delres}</td>
+                                    <td className={`${bc}`}><MdOutlineUpdate className='botton' onClick={()=>sendName(entry.problem_name)}/></td>
+                                    <td className={`${bc}`}><MdDelete className='botton' onClick={()=>handleDelete(entry.problem_name)}/></td>
                                 </tr>
                             )
                         })
